+++ projects/iplab/ngx-color-picker/src/lib/components/chrome-picker/chrome-picker.component.ts
@@ -13,6 +13,7 @@ import {
 import { ColorString } from './../../helpers/color.class';
 import { ColorPickerControl } from './../../helpers/control.class';
 import { getValueByType } from './../../helpers/helper.functions';
+import { Subscription } from 'rxjs';
 
 @Component({
     selector: `chrome-picker`,
@@ -36,6 +37,8 @@ export class ChromePickerComponent implements OnInit, OnChanges, OnDestroy {
 
     @Output()
     public colorChange: EventEmitter<ColorString> = new EventEmitter(false);
+
+    private subscriptions: Array<Subscription> = [];
 
     constructor(private readonly cdr: ChangeDetectorRef) {
     }
@@ -78,15 +81,18 @@ export class ChromePickerComponent implements OnInit, OnChanges, OnDestroy {
                 ]);
         }
 
-        this.control.valueChanges.subscribe((value) => {
-            this.cdr.markForCheck();
-            this.colorChange.emit(getValueByType(value, this.control.initType));
-        });
+        this.subscriptions.push(
+            this.control.valueChanges.subscribe((value) => {
+                this.cdr.markForCheck();
+                this.colorChange.emit(getValueByType(value, this.control.initType));
+            })
+        );
     }
 
     public ngOnDestroy(): void {
-        this.control.unsubscribe();
         this.cdr.detach();
+        this.subscriptions.forEach((subscription) => subscription.unsubscribe());
+        this.subscriptions.length = 0;
     }
 
     public ngOnChanges(changes: SimpleChanges): void {