+++ projects/iplab/ngx-color-picker/src/lib/components/parts/alpha/alpha.component.ts
@@ -5,14 +5,11 @@ import {
     Output,
     ChangeDetectionStrategy,
     OnChanges,
-    Inject,
     ElementRef,
     ViewChild,
-    HostListener,
     Renderer2,
     SimpleChanges
 } from '@angular/core';
-import { DOCUMENT } from '@angular/common';
 import { Color } from './../../../helpers/color.class';
 import { BaseComponent } from './../base.component';
 
@@ -38,14 +35,8 @@ export class AlphaComponent extends BaseComponent implements OnChanges {
 
     private isVertical: boolean = false;
 
-    constructor(renderer: Renderer2, @Inject(DOCUMENT) document, elementRef: ElementRef) {
-        super(document, elementRef, renderer);
-    }
-
-    @HostListener('mousedown', ['$event'])
-    @HostListener('touchstart', ['$event'])
-    public onClick(event: any): void {
-        this.onEventChange(event);
+    constructor(private readonly renderer: Renderer2) {
+        super();
     }
 
     @Input()