+++ projects/iplab/ngx-color-picker/src/lib/components/parts/hue/hue.component.ts
@@ -3,16 +3,13 @@ import {
     ChangeDetectionStrategy,
     Input,
     Renderer2,
-    Inject,
     ElementRef,
     Output,
     EventEmitter,
     ViewChild,
-    HostListener,
     SimpleChanges,
     OnChanges
 } from '@angular/core';
-import { DOCUMENT } from '@angular/common';
 import { Color } from './../../../helpers/color.class';
 import { BaseComponent } from './../base.component';
 
@@ -44,14 +41,8 @@ export class HueComponent extends BaseComponent implements OnChanges {
 
     private isVertical: boolean = false;
 
-    constructor(renderer: Renderer2, @Inject(DOCUMENT) document, elementRef: ElementRef) {
-        super(document, elementRef, renderer);
-    }
-
-    @HostListener('mousedown', ['$event'])
-    @HostListener('touchstart', ['$event'])
-    public onClick(event: any): void {
-        this.onEventChange(event);
+    constructor(private readonly renderer: Renderer2) {
+        super();
     }
 
     @Input()