+++ projects/iplab/ngx-color-picker/src/lib/components/parts/indicator/indicator.component.ts
@@ -1,7 +1,8 @@
-import { Component, Input, HostBinding, HostListener, Renderer2, ElementRef, ChangeDetectionStrategy, Inject } from '@angular/core';
+import { Component, Input, HostBinding, Renderer2, ElementRef, ChangeDetectionStrategy, Inject } from '@angular/core';
 import { DOCUMENT } from '@angular/common';
 import { Color } from './../../../helpers/color.class';
 import { ColorPickerConfig } from './../../../services/color-picker.service';
+import { fromEvent, Subscription } from 'rxjs';
 
 
 @Component({
@@ -25,19 +26,23 @@ export class IndicatorComponent {
         return this.pickerConfig ? this.pickerConfig.indicatorTitle : '';
     }
 
+    private subscriptions: Subscription[] = [];
+
     constructor(
         private readonly pickerConfig: ColorPickerConfig,
         private readonly renderer: Renderer2,
         private readonly elementRef: ElementRef,
         @Inject(DOCUMENT) private readonly document) {
+            this.subscriptions.push(
+                fromEvent(this.elementRef.nativeElement, 'click').subscribe(() => this.onClick())
+            );
     }
 
     public get backgroundColor(): string {
         return this.color.toRgbaString();
     }
 
-    @HostListener('click', ['$event'])
-    public onClick(event: MouseEvent | TouchEvent) {
+    private onClick() {
         const input = this.renderer.createElement('input');
         this.renderer.setStyle(input, 'position', 'absolute');
         this.renderer.setStyle(input, 'top', '-100%');
@@ -48,7 +53,7 @@ export class IndicatorComponent {
                 input.value = this.color.toHslaString();
                 break;
             case 'hex':
-                input.value = this.color.toHexString();
+                input.value = this.color.toHexString(this.color.getRgba().alpha < 1);
                 break;
             default:
                 input.value = this.backgroundColor;