+++ src/app/wrap-examples/wrap-example.module.ts
@@ -1,0 +1,20 @@
+import { NgModule } from '@angular/core';
+import { BrowserModule } from '@angular/platform-browser';
+import { ColorPickerModule } from '@iplab/ngx-color-picker';
+import { ChromeWrapperComponent } from './chrome-picker/chrome-wrapper.component';
+
+
+
+@NgModule({
+  declarations: [
+    ChromeWrapperComponent
+  ],
+  imports: [
+    BrowserModule,
+    ColorPickerModule
+  ],
+  exports: [
+    ChromeWrapperComponent
+  ]
+})
+export class WrapperExampleModule { }