+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+<a name="2.1.0"></a>
+## [2.1.0](https://github.com/pIvan/ngx-color-picker/commit/0c3ecdd853747126e8dea565ccdeea2a9bf7c9f2) (2021-05-26)
+
+### Angular 12 update
+
+<a name="2.0.1"></a>
+## [2.0.1](https://github.com/pIvan/ngx-color-picker/commit/ddbc5869fa74fd64714173ec4dec7a244d649ca1) (2020-5-27)
+
+### hex to string validation fix
+
+<a name="1.0.5"></a>
+## [1.0.5](https://github.com/pIvan/ngx-color-picker/commit/d6b677bef03106d5f2b259e0704aaac1f0b612d3) (2019-10-22)
+
+### Fixed problem with subscriptions when a component is destroy
+### Example of wrapper component added
+
 <a name="1.0.2"></a>
 ## [1.0.2](https://github.com/pIvan/ngx-color-picker/commit/f41102b8e0d725f8804a929c65c7ff9babf2c1fe) (2019-6-10)